backend/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
web3==5.25.0
python-dotenv==0.19.2
pandas==1.4.2